whatpm/t/tree-test-foreign.dat

#data
<math></math>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <math>

#data
<math>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <math>

#data
<math><mi></mi></math>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>

#data
<math><mi>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>

#data
<svg></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg>
|     "x"

#data
<svg>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg>

#data
<svg/>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg>
|     "x"

#data
<svg><textpath></textpath></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg>
|       <textPath>
|     "x"

#data
<svg><foreignobject/></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg>
|       <foreignObject>
|     "x"

#data
<svg><foreignObject></foreignobject></svg>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <svg>
|       <foreignObject>
|     "x"

#data
<svg><foreignObject>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg>
|       <foreignObject>

#data
<svg><foreignobject></body>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <svg>
|       <foreignObject>
|         "x"

#data
<p><svg><foreignobject><address>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <address>

#data
<p><svg><foreignobject><article>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <article>

#data
<p><svg><foreignobject><aside>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <aside>

#data
<p><svg><foreignobject><blockquote>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <blockquote>

#data
<p><svg><foreignobject><center>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <center>

#data
<p><svg><foreignobject><datagrid>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <datagrid>

#data
<p><svg><foreignobject><details>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <details>

#data
<p><svg><foreignobject><dialog>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <dialog>

#data
<p><svg><foreignobject><dir>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <dir>

#data
<p><svg><foreignobject><div>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <div>

#data
<p><svg><foreignobject><dl>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <dl>

#data
<p><svg><foreignobject><figure>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <figure>

#data
<p><svg><foreignobject><footer>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <footer>

#data
<p><svg><foreignobject><h6>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <h6>

#data
<p><svg><foreignobject><header>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <header>

#data
<p><svg><foreignobject><menu>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <menu>

#data
<p><svg><foreignobject><nav>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <nav>

#data
<p><svg><foreignobject><ol>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <ol>

#data
<p><svg><foreignobject><p>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <p>

#data
<p><svg><foreignobject><section>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <section>

#data
<p><svg><foreignobject><ul>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <ul>

#data
<p><svg><foreignobject><pre>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <pre>

#data
<p><svg><foreignobject><listing>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <listing>

#data
<p><svg><foreignobject><form>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <form>

#data
<p><svg><foreignobject><fieldset>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <fieldset>

#data
<p><svg><foreignobject><li>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <li>

#data
<p><svg><foreignobject><dd>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <dd>

#data
<p><svg><foreignobject><dt>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <dt>

#data
<p><svg><foreignobject><plaintext>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <plaintext>

#data
<address><svg><foreignobject></address>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <address>
|       <svg>
|         <foreignObject>
|           "x"

#data
<article><svg><foreignobject></article>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <svg>
|         <foreignObject>
|           "x"

#data
<aside><svg><foreignobject></aside>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <svg>
|         <foreignObject>
|           "x"

#data
<ul><svg><foreignobject></ul>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <svg>
|         <foreignObject>
|           "x"

#data
<form><svg><foreignobject></form>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <svg>
|         <foreignObject>
|           "x"

#data
<p><svg><foreignobject></p>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <p>
|           "x"

#data
<dd><svg><foreignobject></dd>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <dd>
|       <svg>
|         <foreignObject>
|           "x"

#data
<dt><svg><foreignobject></dt>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <dt>
|       <svg>
|         <foreignObject>
|           "x"

#data
<li><svg><foreignobject></li>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <li>
|       <svg>
|         <foreignObject>
|           "x"

#data
<h1><svg><foreignobject></h1>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <h1>
|       <svg>
|         <foreignObject>
|           "x"

#data
<nobr><svg><foreignobject><nobr></nobr></foreignObject></svg></nobr>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nobr>
|       <svg>
|         <foreignObject>
|           <nobr>

#data
<button><svg><foreignobject><button></button></foreignObject></svg></button>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <svg>
|         <foreignObject>
|           <button>

#data
<applet><svg><foreignobject><applet></applet></foreignObject></svg></applet>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <svg>
|         <foreignObject>
|           <applet>

#data
<applet><svg><foreignobject></applet>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <svg>
|         <foreignObject>
|           "x"

#data
<button><svg><foreignobject></button>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <svg>
|         <foreignObject>
|           "x"

#data
<marquee><svg><foreignobject></marquee>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <svg>
|         <foreignObject>
|           "x"

#data
<object><svg><foreignobject></object>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <object>
|       <svg>
|         <foreignObject>
|           "x"

#data
<p><svg><foreignobject><table></table>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <table>

#data
<p><svg><foreignobject><hr>
#errors
no DOCTYPE
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <svg>
|         <foreignObject>
|           <hr>

#data
<option><svg><foreignobject></option>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <option>
|       <svg>
|         <foreignObject>
|           "x"

#data
<optgroup><svg><foreignobject></optgroup>x
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
#document
| <html>
|   <head>
|   <body>
|     <optgroup>
|       <svg>
|         <foreignObject>
|           "x"

#data
<ruby><rt><svg><foreignobject><rp>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rt>
|         <svg>
|           <foreignObject>
|             <rp>

#data
<ruby><rt><svg><foreignobject><rt>
#errors
no DOCTYPE
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rt>
|         <svg>
|           <foreignObject>
|             <rt>
